Use relay service token endpoint
